+++ packages/moonstone/Spinner/Spinner.js
@@ -58,10 +58,14 @@ const SpinnerCore = kind({
 
 	render: ({children, css, ...rest}) => (
 		<div aria-live="off" role="alert" {...rest}>
-			<div className={css.ballDecorator}>
-				<div className={`${css.ball} ${css.ball1}`} />
-				<div className={`${css.ball} ${css.ball2}`} />
-				<div className={`${css.ball} ${css.ball3}`} />
+			<div className={css.bg}>
+				<div className={css.decorator}>
+					<div className={css.fan1} />
+					<div className={css.fan2} />
+					<div className={css.fan3} />
+					<div className={css.fan4} />
+					<div className={css.cap} />
+				</div>
 			</div>
 			{children ?
 				<Marquee className={css.client} marqueeOn="render" alignment="center">