+++ packages/ui/Repeater/Repeater.js
@@ -5,9 +5,10 @@
  * @exports Repeater
  */
 
-import React from 'react';
-import PropTypes from 'prop-types';
 import kind from '@enact/core/kind';
+import EnactPropTypes from '@enact/core/internal/prop-types';
+import PropTypes from 'prop-types';
+import React from 'react';
 
 /**
  * A stateless component that stamps out copies of `childComponent`.
@@ -26,11 +27,11 @@ const RepeaterBase = kind({
 		 *
 		 * This can be a React component or a string describing a DOM node (e.g. `'div'`).
 		 *
-		 * @type {Component}
+		 * @type {String|Component}
 		 * @required
 		 * @public
 		 */
-		childComponent: PropTypes.oneOfType([PropTypes.string, PropTypes.func]).isRequired,
+		childComponent: EnactPropTypes.renderable.isRequired,
 
 		/**
 		 * An array of data to be mapped onto the `childComponent`.
@@ -65,6 +66,17 @@ const RepeaterBase = kind({
 		childProp: PropTypes.string,
 
 		/**
+		 * Component type to wrap around all of the repeated elements.
+		 *
+		 * This can be a string describing a DOM node or React component (e.g. `'div'` or `Layout`).
+		 *
+		 * @type {String|Component}
+		 * @default 'span'
+		 * @public
+		 */
+		component: EnactPropTypes.renderable,
+
+		/**
 		 * The property on each `childComponent` that receives the index of the item in the `Repeater`.
 		 *
 		 * @type {String}
@@ -83,8 +95,9 @@ const RepeaterBase = kind({
 	},
 
 	defaultProps: {
-		indexProp: 'data-index',
-		childProp: 'children'
+		childProp: 'children',
+		component: 'span',
+		indexProp: 'data-index'
 	},
 
 	computed: {
@@ -103,13 +116,13 @@ const RepeaterBase = kind({
 		}
 	},
 
-	render: (props) => {
-		delete props.childComponent;
-		delete props.childProp;
-		delete props.indexProp;
-		delete props.itemProps;
+	render: ({component: Component, ...rest}) => {
+		delete rest.childComponent;
+		delete rest.childProp;
+		delete rest.indexProp;
+		delete rest.itemProps;
 
-		return <span role="list" {...props} />;
+		return <Component role="list" {...rest} />;
 	}
 });